requirements.txt
valkey
matplotlib
numpy
scipy
boto3
requests
requests-aws4auth
python-snappy
protobuf
grpcio-tools
